pyproject.toml
[tool.poetry]
name = "kolory"
version = "0.1.0"
description = ""
authors = ["Ruslan Sirazhetdinov <ruslansir@icloud.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
numpy = "^2.1.2"
scikit-learn = "^1.5.2"
matplotlib = "^3.9.2"
opencv-python = "^4.10.0.84"
plotly = "^5.24.1"
pillow = "<11"
streamlit = "^1.39.0"
watchdog = "^5.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
